Moves Conquer_Build_Fading_Table to new fading.c file.
