+++ common/fading.c
@@ -88,4 +88,82 @@ void* Build_Fading_Table(void const* palette, void* dest, int color, int frac)
     return dest;
 }
 
+void* Conquer_Build_Fading_Table(const void* palette, void* dest, int color, int frac)
+{
+    const int ALLOWED_COUNT = 16;
+    const int ALLOWED_START = 256 - ALLOWED_COUNT;
+
+    if (!palette || !dest) {
+        return 0;
+    }
+
+    const unsigned char* pal = palette;
+    unsigned char* dst = dest;
+
+    if ((unsigned)frac >= 256) {
+        frac = 255;
+    }
+
+    int fraction = frac >> 1;
+    unsigned palindex = color * 3;
+    unsigned char targetred = pal[palindex++];
+    unsigned char targetgreen = pal[palindex++];
+    unsigned char targetblue = pal[palindex];
+    unsigned tableindex = 0;
+    dst[0] = 0;
+
+    // Remap most pal entries to the last 16 entries that are the most faded colours.
+    for (int i = 1; i < ALLOWED_START; ++i) {
+        // Decide what the "perfect" match would be for our adjusted color.
+        palindex = i * 3;
+        signed char original = pal[palindex++];
+        signed short tmp = ((original - targetred) * fraction) << 1;
+        unsigned char idealred = original - (tmp >> 8);
+        original = pal[palindex++];
+        tmp = ((original - targetgreen) * fraction) << 1;
+        unsigned char idealgreen = original - (tmp >> 8);
+        original = pal[palindex];
+        tmp = ((original - targetblue) * fraction) << 1;
+        unsigned char idealblue = original - (tmp >> 8);
+
+        const unsigned char* fade = pal + ALLOWED_START * 3;
+        unsigned matchcolor = color;
+        unsigned matchvalue = (unsigned)(-1);
+        unsigned matchindex = ALLOWED_START;
+
+        // Find the closest match that actually exists in the allowed palette range for the adjusted color.
+        for (int j = 0; j < ALLOWED_COUNT; ++j) {
+            // Build a distance of our current color to the ideal faded color.
+            signed char diff = *fade++ - idealred;
+            unsigned value = diff * diff;
+            diff = *fade++ - idealgreen;
+            value += diff * diff;
+            diff = *fade++ - idealblue;
+            value += diff * diff;
+
+            // In the event of a tie, this will match the later color in the palette.
+            if (value < matchvalue) {
+                matchvalue = value;
+                matchcolor = matchindex;
+            }
+
+            // Perfect match, use it.
+            if (value == 0) {
+                break;
+            }
+
+            ++matchindex;
+        }
+
+        dst[i] = matchcolor;
+    }
+
+    // Make sure last 16 values just remap to themselves.
+    for (int i = ALLOWED_START; i < 256; ++i) {
+        dst[i] = i;
+    }
+
+    return dest;
+}
+
 #endif
