+++ common/fading.h
@@ -17,6 +17,7 @@ extern "C" {
 #endif
 
 void* Build_Fading_Table(void const* palette, void* dest, int color, int frac);
+void* Conquer_Build_Fading_Table(void const* palette, void* dest, int color, int frac);
 
 #ifdef __cplusplus
 }
