+++ common/miscasm.h
@@ -15,7 +15,6 @@ int First_True_Bit(void const* array);
 int First_False_Bit(void const* array);
 int _Bound(int original, int min, int max);
 #define Bound _Bound
-void* Conquer_Build_Fading_Table(void const* palette, void* dest, int color, int frac);
 int Reverse_Long(int number);
 void strtrim(char* buffer);
