+++ tests/fading.cpp
@@ -1,4 +1,4 @@
-#include "common/miscasm.h"
+#include "common/fading.h"
 
 #include <stdint.h>
 #include <iostream>
